Add grace period for fast OOB ack + resending rqst
Allow grace period when receiving a fast OOB ack:
A boot server can take more than 250ms to reply, especially to a pairing
request. When the boot server replies with a fast OOB ack (CTL and ACK bits
on, zero payload length), the boot loader is informed that its request is being
processed. It can therefore introduce a grace period by extending the 250ms
timeout to 2.5secs. If no such fast OOB ack is received, the original timeout
logic is kept, maintaining backwards compatibility.

Send upgrade request after firmware download:
We re-send the upgrade request packet following a successful firmware download
to indicate to the boot server that we're now running the target firmware. This
allows the boot server to update the settings for this specific node.